Remove the improved AI Blender client and related documentation; transition to a simpler AI-integrated Blender client using official SDKs (anthropic, openai). Update test scripts to reflect the removal of the improved client and comparison functionality.
