--- IMPROVED_AI_INTEGRATION.md
@@ -1,214 +1,0 @@
-# 🧠 Improved AI Integration with Proper SDKs
-
-## 🎯 **Problem Solved**
-
-**Original Question:** "Should we be connecting to Claude and OpenAI directly through httpx.AsyncClient? Is there no service/opensource project that handles this for us?"
-
-**Answer:** **NO!** Raw HTTP calls are not recommended. There are much better alternatives.
-
-## ✅ **Solution: Proper AI SDKs**
-
-We've implemented **multiple approaches** using battle-tested libraries instead of raw HTTP:
-
-### **1. Official SDKs (Recommended)**
-- **Anthropic SDK** (`anthropic`) for Claude
-- **OpenAI SDK** (`openai`) for GPT models
-
-### **2. Unified Interfaces**
-- **LiteLLM** - Single interface for 100+ AI providers
-- **LangChain** - Full AI framework (future enhancement)
-
-## 🚀 **What's Been Implemented**
-
-### **New Files Created:**
-- `ai_blender_client_improved.py` - Improved client with proper SDKs
-- `AI_SDK_COMPARISON.md` - Detailed comparison of approaches
-- Updated `pyproject.toml` with proper dependencies
-- Enhanced `scripts/test_blender.sh` with new commands
-
-### **New Dependencies Added:**
-```bash
-uv add anthropic openai litellm
-```
-
-### **New Commands Available:**
-```bash
-./scripts/test_blender.sh ai-improved   # Use improved client
-./scripts/test_blender.sh compare       # Compare SDK approaches
-```
-
-## 📊 **Before vs After Comparison**
-
-### **❌ Previous Approach (Raw HTTP)**
-```python
-# Don't do this!
-async with httpx.AsyncClient() as client:
-    response = await client.post(
-        "https://api.anthropic.com/v1/messages",
-        headers={"Authorization": f"Bearer {api_key}"},
-        json={
-            "model": "claude-3-5-sonnet-20241022",
-            "max_tokens": 4000,
-            "messages": [{"role": "user", "content": message}]
-        }
-    )
-    result = response.json()
-    return result["content"][0]["text"]
-```
-
-**Problems:**
-- ❌ Manual error handling
-- ❌ No retries or rate limiting
-- ❌ Complex authentication
-- ❌ No type safety
-- ❌ Parsing overhead
-
-### **✅ Improved Approach (Official SDK)**
-```python
-# Much better!
-import anthropic
-
-client = anthropic.AsyncAnthropic(api_key=api_key)
-response = await client.messages.create(
-    model="claude-3-5-sonnet-20241022",
-    max_tokens=4000,
-    messages=[{"role": "user", "content": message}]
-)
-return response.content[0].text
-```
-
-**Benefits:**
-- ✅ **60% less code**
-- ✅ Built-in error handling & retries
-- ✅ Automatic rate limiting
-- ✅ Type safety
-- ✅ Streaming support
-- ✅ Better debugging
-
-## 🎛️ **Multiple Implementation Approaches**
-
-### **Approach 1: Official SDKs**
-```python
-# Choose your provider
-client = AIBlenderClientImproved(
-    ai_provider="claude",      # or "openai"
-    approach="official"
-)
-```
-
-### **Approach 2: LiteLLM (Unified)**
-```python
-# Same interface for all providers
-client = AIBlenderClientImproved(
-    ai_provider="claude",      # or "openai", "gemini", etc.
-    approach="litellm"
-)
-```
-
-## 🧪 **Testing Results**
-
-```bash
-$ ./scripts/test_blender.sh compare
-🔬 COMPARISON: Different AI Integration Approaches
-============================================================
-
-📊 Official SDKs (anthropic, openai)
-   Approach: official
-   ✅ Blender connection: OK
-   ✅ AI client setup: OK
-
-📊 LiteLLM (unified interface)  
-   Approach: litellm
-   ✅ Blender connection: OK
-   ✅ AI client setup: OK
-```
-
-## 📋 **Usage Guide**
-
-### **1. Install Dependencies**
-```bash
-uv add anthropic openai litellm
-```
-
-### **2. Set Environment Variables**
-```bash
-# Copy template
-cp env_template.txt .env
-
-# Edit .env with real API keys
-ANTHROPIC_API_KEY=sk-ant-your-key-here
-OPENAI_API_KEY=sk-your-openai-key-here
-```
-
-### **3. Start HTTP Server**
-```bash
-# Terminal 1: Start Blender MCP HTTP server
-./scripts/test_blender.sh http
-```
-
-### **4. Test Different Approaches**
-```bash
-# Terminal 2: Compare approaches
-./scripts/test_blender.sh compare
-
-# Use improved client (default: Claude + Official SDK)
-./scripts/test_blender.sh ai-improved
-
-# Use OpenAI instead
-AI_PROVIDER=openai ./scripts/test_blender.sh ai-improved
-
-# Use LiteLLM unified interface
-AI_APPROACH=litellm ./scripts/test_blender.sh ai-improved
-```
-
-## 🎯 **Key Advantages of New Implementation**
-
-### **1. Reliability**
-- Official SDKs handle edge cases
-- Built-in retry logic
-- Proper timeout handling
-
-### **2. Maintainability**
-- No manual HTTP management
-- Official support and updates
-- Type safety and intellisense
-
-### **3. Flexibility**
-- Easy provider switching
-- Multiple implementation approaches
-- Future-proof architecture
-
-### **4. Features**
-- Streaming support (when needed)
-- Rate limiting built-in
-- Better error messages
-
-## 🔮 **Future Enhancements**
-
-### **Already Available:**
-- ✅ Official SDKs (anthropic, openai)
-- ✅ LiteLLM unified interface
-- ✅ Environment-based configuration
-
-### **Potential Additions:**
-- 🔄 LangChain integration for complex AI workflows
-- 📊 Cost tracking and usage analytics
-- 🔄 Provider fallback chains
-- 🤖 AI agent patterns for complex Blender tasks
-
-## 📚 **Documentation**
-
-- `AI_SDK_COMPARISON.md` - Detailed comparison of all approaches
-- `ai_blender_client_improved.py` - Implementation with proper SDKs
-- `AI_INTEGRATION.md` - Original HTTP-based implementation (legacy)
-
-## 🎉 **Conclusion**
-
-**The answer to your question:** Yes, there are excellent open-source projects that handle AI API connections much better than raw HTTP calls!
-
-**We now use:**
-1. **Official SDKs** - `anthropic` and `openai` packages
-2. **LiteLLM** - Unified interface for 100+ providers
-3. **Proper architecture** - Clean, maintainable, and extensible
-
-This is a **significant improvement** over raw `httpx.AsyncClient` calls and follows industry best practices.
